+++ src/main/java/com/furkanerguldurenler/services/IShoppingListService.java
@@ -4,7 +4,7 @@ import com.furkanerguldurenler.dto.AddIngredientsRequestDto;
 
 public interface IShoppingListService {
 
-	public void addIngredients(AddIngredientsRequestDto addIngredientsRequestDto);
+    public void addIngredients(AddIngredientsRequestDto addIngredientsRequestDto);
 
-	public void removeIngredientById(Integer userId, Integer ingredientId);
+    public void removeIngredientById(Long userId, Integer ingredientId);
 }
